pyproject.toml
[tool.poetry]
name = "clickup-api"
version = "0.1.0"
description = "MIT"
authors = ["venysssssssssss <shablaur7@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
openpyxl = "^3.1.2"
fastapi = "^0.111.0"
blue = "^0.9.1"
isort = "^5.13.2"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
pyngrok = "^7.1.6"
uvicorn = { version = "^0.29.0", python = ">=3.8,<4.0" }
pandas = "^2.2.2"
duckdb = "^0.10.3"
fastapi-cache = "^0.1.0"
fastapi-cache2 = "^0.2.1"
redis = "^5.0.4"
aioredis = "1.3.1"
msgpack-python = "^0.5.6"
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.27"
mkdocstrings = "^0.25.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
